grafana/otel-jaegar/requirements.txt
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp-proto-http 
opentelemetry-exporter-otlp
opentelemetry-instrumentation-flask
flask
mysql-connector-python